Both VN sensors reading w/o issue

+++ libraries/AP_Airspeed/AP_Airspeed_SST_ND.h
@@ -53,7 +53,7 @@ private:
     bool matchModel(uint8_t* reading);
     bool probe(uint8_t bus, uint8_t address);
     void _collect();
-    float _get_pressure(int16_t dp_raw) const;
+    float _get_pressure(uint32_t dp_raw) const;
     float _get_temperature(int8_t dT_int, int8_t dT_frac) const;
     bool range_change_needed(float last_pressure);
     void update_range(void);
